requirements.txt
alembic==0.8.9
aniso8601==1.2.0
click==6.6
coverage==4.3.1
Flask==0.12
Flask-Login==0.4.0
Flask-Migrate==2.0.2
Flask-MySQL==1.4.0
flask-paginate==0.4.5
Flask-RESTful==0.3.5
Flask-Script==2.0.5
Flask-SQLAlchemy==2.1
Flask-Testing==0.6.1
Flask-WTF==0.13.1
itsdangerous==0.24
Jinja2==2.8.1
Mako==1.0.6
MarkupSafe==0.23
passlib==1.7.0
pluggy==0.4.0
py==1.4.32
pycrypto==2.6.1
PyJWT==1.4.2
PyMySQL==0.7.9
python-dateutil==2.6.0
python-editor==1.0.3
pytz==2016.10
six==1.10.0
SQLAlchemy==1.1.4
SQLAlchemy-Utils==0.32.12
tox==2.5.0
virtualenv==15.1.0
Werkzeug==0.11.15
WTForms==2.1
